healthd-ext: Fix the bootup delay issue
On some platforms, the fuel gauge is being implemented on the APPS side
with 3rd party fuel gauges like bq27421. These targets may see
unnecessary delay in booting up due to the 10s timeout in the
qti_healthd_board_init(), where we wait for the 'battery/capacity'
node to be created, to make sure health service starts after the charger
firmware is up.

So add a change to bailout early if the 'bq27421-0/capacity' is present.

Change-Id: I2c3ff989d3daf93d28d9bf948c24758a78e52a86